test(#4878): throwaway git-mv rename verification, do not merge
Pure git mv of src/review/inline-finding-category-parse.ts (zero
content change) plus the two import-path updates it requires, to
observe how Codecov represents a rename versus a delete+add. This
commit is never meant to land -- see #4878.
